+++ tests/test_docs.py
@@ -2,15 +2,15 @@
 
 from __future__ import absolute_import, print_function, unicode_literals
 
-from django.test import TestCase
-
 from django_ftl import activate
 from django_ftl.bundles import Bundle
+
+from .base import TestBase
 
 ftl_bundle = Bundle(['tests/docs.ftl'])
 
 
-class TestDocs(TestCase):
+class TestDocs(TestBase):
 
     def test_usage_docs(self):
         # These tests parallel the code in the usage.rst docs