+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
 * Changed development autoreload mechanism to not interfere with Django's
   development server autoreload.
 * Bug fix for case when invalid mode is specified in template tag.
+* Various fixes and improvements to middlewares (plus tests)
 
 0.9 (2018-09-10)
 ++++++++++++++++